Componentize DocumentView
