+++ frontend/src/pages/workspace/document/Index.tsx
@@ -1,7 +1,6 @@
 import { useEffect } from "react";
-import Editor from "../../../components/editor/Editor";
 import * as yorkie from "yorkie-js-sdk";
-import { selectEditor, setClient, setDoc } from "../../../store/editorSlice";
+import { setClient, setDoc } from "../../../store/editorSlice";
 import { useDispatch, useSelector } from "react-redux";
 import {
 	YorkieCodeMirrorDocType,
@@ -9,21 +8,17 @@ import {
 } from "../../../utils/yorkie/yorkieSync";
 import randomColor from "randomcolor";
 import Color from "color";
-import { Box, Paper } from "@mui/material";
-import Resizable from "react-resizable-layout";
-import { useWindowWidth } from "@react-hook/window-size";
-import Preview from "../../../components/editor/Preview";
+import { Box } from "@mui/material";
 import { useParams } from "react-router-dom";
 import { selectUser } from "../../../store/userSlice";
 import { useGetDocumentQuery } from "../../../hooks/api/workspaceDocument";
 import { useGetWorkspaceQuery } from "../../../hooks/api/workspace";
+import DocumentView from "../../../components/editor/DocumentView";
 
 function DocumentIndex() {
 	const dispatch = useDispatch();
 	const params = useParams();
 	const userStore = useSelector(selectUser);
-	const windowWidth = useWindowWidth();
-	const editorStore = useSelector(selectEditor);
 	const { data: workspace } = useGetWorkspaceQuery(params.workspaceSlug);
 	const { data: document } = useGetDocumentQuery(workspace?.id, params.documentId);
 
@@ -71,51 +66,7 @@ function DocumentIndex() {
 
 	return (
 		<Box height="calc(100% - 64px)">
-			{/* For Markdown Preview Theme */}
-			<div className="wmde-markdown-var" />
-			{editorStore.mode === "both" && (
-				<Resizable axis={"x"} initial={windowWidth / 2} min={400}>
-					{({ position: width, separatorProps }) => (
-						<div
-							id="wrapper"
-							style={{
-								display: "flex",
-								height: "100%",
-								overflow: "hidden",
-							}}
-						>
-							<div id="left-block" style={{ width }}>
-								<Editor />
-							</div>
-							<Paper
-								id="splitter"
-								{...separatorProps}
-								sx={{
-									height: "100%",
-									width: 8,
-									borderRadius: 0,
-									cursor: "col-resize",
-									zIndex: 100,
-								}}
-							/>
-							<div
-								className="right-block"
-								style={{ width: `calc(100% - ${width}px)`, overflow: "auto" }}
-							>
-								<Box sx={{ p: 4 }} height="100%">
-									<Preview />
-								</Box>
-							</div>
-						</div>
-					)}
-				</Resizable>
-			)}
-			{editorStore.mode === "read" && (
-				<Box sx={{ p: 4, overflow: "auto" }} height="100%">
-					<Preview />
-				</Box>
-			)}
-			{editorStore.mode === "edit" && <Editor />}
+			<DocumentView />
 		</Box>
 	);
 }
