+++ frontend/src/pages/workspace/document/share/Index.tsx
@@ -1,7 +1,12 @@
 import { Box } from "@mui/material";
+import DocumentView from "../../../../components/editor/DocumentView";
 
 function DocumentShareIndex() {
-	return <Box></Box>;
+	return (
+		<Box height="calc(100% - 64px)">
+			<DocumentView />
+		</Box>
+	);
 }
 
 export default DocumentShareIndex;
